Move pytest_terminal_summary hook to matcher
